+++ src/app.py
@@ -11,7 +11,7 @@ from sys import exit
 
 config_path = os.environ["HUB_CONFIG_FILE"]
 
-version = "v1.12.5"
+version = "v1.12.6"
 
 for argv in sys.argv:
     if argv.endswith(".py"):